django/requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-bootstrap-modal-forms==2.1.0
django-cors-headers==3.7.0
django-phonenumber-field==5.0.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.10
oauthlib==3.1.0
packaging==20.9
phonenumbers==8.12.19
Pillow==8.1.2
pip==21.0.1
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools==54.1.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3


python-dateutil~=2.8.1